What Do We See Here ?
Imagine if you have 2 options in an exam and there are 50 questions. Normally, we can expect to give 25 answers correctly, just by luck. We may get a result around 20 or 30 correct answers. But if you try it on 500 questions, you will get closer to 250 than if you try 50 times and get closer to 25.
I wanted to calculate this reality of the calculations of luck with math. Try to set same number of options but with more trying numbers. Most probably, you will see closer ratio to the ratio it should be. We see that if you try more, you get closer results to reality. Keep trying.